+++ circumflex-core/src/main/scala/exception.scala
@@ -4,14 +4,13 @@ package ru.circumflex.core
 
 All exceptions thrown from Circumflex components should extend `CircumflexException` class.
 */
-
-/**
- * Indicates a common error occured inside Circumflex component.
- *
- * For more information refer to
- * <a href="http://circumflex.ru/api/2.0.2/circumflex-core/exception.scala">exception.scala</a>.
- */
 class CircumflexException(msg: String, cause: Throwable = null)
     extends RuntimeException(msg, cause) {
   def this(cause: Throwable) = this(null, cause)
+}
+
+class ValidationException(val errors: Seq[Msg])
+    extends CircumflexException("Validation failed.") {
+  def this(msg: Msg) = this(List(msg))
+  def this(msg: Msg, msgs: Msg*) = this(List(msg) ++ msgs.toSeq)
 }