+++ circumflex-markeven/src/main/scala/util.scala
@@ -12,15 +12,11 @@ class Protector {
   protected var protectHash: Map[String, CharSequence] = Map()
   protected var unprotectHash: Map[CharSequence, String] = Map()
 
-  /**
-   * Generates a random hash key.
-   */
+
   def randomKey = "!}" + (0 until keySize).foldLeft("")((s, i) =>
     s + chars.charAt(rnd.nextInt(chars.length)))
 
-  /**
-   * Adds the specified token to hash and returns the protection key.
-   */
+
   def addToken(t: CharSequence): String = unprotectHash.get(t) match {
     case Some(key) => key
     case _ =>
@@ -30,19 +26,13 @@ class Protector {
       key
   }
 
-  /**
-   * Attempts to retrieve an encoded sequence by specified `key`.
-   */
+
   def decode(key: String): Option[CharSequence] = protectHash.get(key)
 
-  /**
-   * Returns hash keys that are currently in use.
-   */
+
   def keys = protectHash.keys
 
-  /**
-   * Clears both hashes.
-   */
+
   def clear: Unit = {
     protectHash = Map()
     unprotectHash = Map()
@@ -59,7 +49,7 @@ The rationale behind low-level replacements is that `java.util.regex.Matcher` su
 `StringBuffer` in `appendReplacement` and `appendTail` methods. Since our processing environment
 does not support multithreading, we avoid synchronization costs using `StringBuilder` instead.
 */
-class StringEx(var buffer: StringBuilder) extends CharSequence {
+class StringEx(var buffer: StringBuilder) extends CharSequence with Cloneable {
   def this(cs: CharSequence) = this(new StringBuilder(cs))
 
   // A tricky one: uses specified `pattern` to point the start of replacement and lets
@@ -80,24 +70,17 @@ class StringEx(var buffer: StringBuilder) extends CharSequence {
   }
 
   def replaceAll(pattern: Pattern, replacement: Matcher => CharSequence): this.type = {
-    replaceIfFound(pattern, replacement)
-    return this
-  }
-
-  def replaceIfFound(pattern: Pattern, replacement: Matcher => CharSequence): Boolean = {
     var lastIndex = 0;
     val m = pattern.matcher(buffer)
     val sb = new StringBuilder(buffer.length)
-    var result = false
     while (m.find()) {
       sb.append(buffer.subSequence(lastIndex, m.start))
       sb.append(replacement(m))
       lastIndex = m.end
-      result = true;
     }
     sb.append(buffer.subSequence(lastIndex, buffer.length))
     this.buffer = sb
-    return result
+    return this
   }
 
   def replaceFirst(pattern: Pattern, replacement: Matcher => CharSequence): Int = {
@@ -144,6 +127,11 @@ class StringEx(var buffer: StringBuilder) extends CharSequence {
   }
 
   def outdent(): this.type = replaceAll(regexes.outdent(4), "")
+
+  def substring(start: Int, end: Int = length): this.type = {
+    buffer = new StringBuilder(subSequence(start, end))
+    return this
+  }
 
   def startsWith(cs: CharSequence): Boolean = {
     if (cs.length == 0) return true
@@ -206,6 +194,8 @@ class StringEx(var buffer: StringBuilder) extends CharSequence {
     if (m.matches) Some(function(m))
     else None
   }
+
+  override def clone: StringEx = super.clone.asInstanceOf[StringEx]
 
   // inherited from `CharSequence`