+++ circumflex-orm/src/main/scala/exception.scala
@@ -5,13 +5,6 @@ package ru.circumflex.orm
 The `ORMException`s are thrown by Circumflex ORM components and usually mean that current
 transaction needs to be rolled back and closed ASAP.
 */
-
-/**
- * General exception in Circumflex ORM.
- *
- * For more information refer to
- * <a href="http://circumflex.ru/api/2.0.2/circumflex-orm/exception.scala">exception.scala</a>.
- */
 class ORMException(msg: String, cause: Throwable) extends Exception(msg, cause) {
   def this(msg: String) = this(msg, null)
   def this(cause: Throwable) = this(null, cause)