+++ circumflex-web/src/main/scala/cookie.scala
@@ -12,13 +12,6 @@ is often referred to as _cookie_.
 Circumflex Web Framework helps you set response cookies and access request cookies
 throughout your application using case class `HttpCookie`. It is mutable and dead simple.
 */
-
-/**
- * Provides convenient mutable wrapper to work with HTTP cookies.
- *
- * For more information refer to
- * <a href="http://circumflex.ru/api/2.0.2/circumflex-web/cookie.scala">cookie.scala</a>.
- */
 case class HttpCookie(var name: String,
                       var value: String,
                       var domain: String = null,
@@ -44,13 +37,6 @@ supplying `javax.servlet.Cookie` as an argument to `apply` method of `HttpCookie
 
     HttpCookie(rawCookie)
 */
-
-/**
- * Converts the `javax.serlvet.Cookie` into `HttpCookie`.
- *
- * For more information refer to
- * <a href="http://circumflex.ru/api/2.0.2/circumflex-web/cookie.scala">cookie.scala</a>.
- */
 object HttpCookie {
   def apply(cookie: Cookie): HttpCookie =
     new HttpCookie(